// src/interface/weeklyreport/type.ts
export interface InvalidTimesheet {
  last_attendance_attempt: string;
  first_attendance_attempt: string;
  attendance_missing_from: string;
  attendance_missing_to: string;
  id: number;
  reason: boolean;
  remarks: string;
  validated: string;
  employee_id: number;
  employee_name: string;
  employee_code: string;
  department: string;
  position: boolean;
  invalid_date: string;
  invalid_type: string;
  shift_from: string;
  shift_to: string;
  shift_break: boolean;
  real_time_attendance_data: string;
  validation_data: string;
  content: boolean;
  create_date: string;
  keep_in_time: boolean;
}